Improve coutning items assigned to a taxonomy
The initial way of counting the items in taxonomy term list template
relied on a map of a taxonomy terms to the count of items
assigned to them being passed to the themplate. This approach was chosen
to make the template less project-specific - that way, it could be used
for listing item count per taxonomy also for items that aren't stored in
the posts field of a taxonomy term object. It was done to make forking
a general Statik theme or a Statik starter project out of this one easier.

However, for this project, all taxonomies would be assigneable only
to posts, therefore the map would always be a mapping of taxonomy terms
to posts. Making it for each view would require some repetitive code.

This commit fixes that by abandoning unnecessary (for now) genericness
of the counting mechanism and making the template rely on the items
being stored in the posts field.