Make footer content part of the base template
This change simplifies handling footer. Instead of being read from
a separate markdown file and using MarkdownPartial model, the content
of footer is now hardcoded into base template.

This is one of changes inteded to simplify the project by removing
some models and reliance on them when including content that does
not benefit significantly from being created and maintained as separate
markdown files. This makes the "theme" used by the project less flexible,
but it shouldn't be a problem for two reasons:

* due to the way Statik works, the "theme" is not really a theme - it's
just a part of the project, even though with some effort it could be
maintained separately
* even with hardcoded, site-specific elements, templates will be simpler
to adjust to other, similar project, just because they rely on fewer
Statik models.