Add a "deslugify" filter
The filter will be used to turn names of model instances (which have
slug-like syntax) into strings that can be used as titles in on rendered
pages, with the first letter capitalized and the hyphens replaced with
whitespace.

This doesn't take into account cases of hyphen being actually a part of
a title, but will do for now.