Make navigation generation use deslugify and pk
This commit replaces previous way of generating contents of non-page
links in navigation menu (that is: taking the name of the view and
capitalizing it) with a new one: taking .pk property of  TopMenuItem
instance and deslugifying it.

This fixes contents of elements consisting of more than one word
and makes them depend on a name of a file that contains definition
of a TopMenuItem instance.