Refactor top menu
The menu is being refactored by making it rely on a list of views
and a list of pages included in it, both declared as static variables
in configuration file, instead of relying on a dynamically selected
list of instances of TopMenuItem.

TopMenuItem and its instances are no longer needed, so they are being
removed.